cloud_functions/cf-i1-train/requirements.txt
google-cloud-logging==3.9.0
google-cloud-aiplatform==1.100.0
google-cloud-storage==2.11.0
PyYAML==6.0.1